requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asarPy==1.0.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
Cython==3.1.1
dotenv==0.9.9
fastapi==0.115.0
Flask==3.1.0
google-auth==2.42.1
google-genai==1.48.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10.dev0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==25.0
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pydantic==2.9.2
pydantic-settings==2.3.4
pydantic_core==2.23.4
Pygments==2.19.1
PyGObject==3.50.1
python-dotenv==1.2.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.46.1
